% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqp.R
\name{get_dims}
\alias{get_dims}
\title{Get dimensions of a PIQP model}
\usage{
get_dims(model, ...)
}
\arguments{
\item{model}{A \code{piqp_model} object}

\item{...}{not used}
}
\value{
a list with named elements \code{n}, \code{p}, and \code{m}
}
\description{
Get dimensions of a PIQP model
}
