% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqp.R
\name{get_settings}
\alias{get_settings}
\title{Get settings of a PIQP model}
\usage{
get_settings(model, ...)
}
\arguments{
\item{model}{A \code{piqp_model} object}

\item{...}{not used}
}
\value{
a list of current settings
}
\description{
Get settings of a PIQP model
}
