% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{make_csc_matrix}
\alias{make_csc_matrix}
\title{Convert a plain matrix or simple triplet form matrix to a \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix} (implicit) form}
\usage{
make_csc_matrix(x)
}
\arguments{
\item{x}{a matrix or a simple triplet form matrix}
}
\value{
a list of row pointer, column pointer, and values corresponding to a \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix} object
}
\description{
Convert a plain matrix or simple triplet form matrix to a \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix} (implicit) form
}
