% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{make_csc_symm_matrix}
\alias{make_csc_symm_matrix}
\title{Convert a plain matrix or simple triplet form matrix to a \link[Matrix:dsCMatrix-class]{Matrix::dsCMatrix} (implicit, upper) form}
\usage{
make_csc_symm_matrix(m)
}
\arguments{
\item{m}{a matrix or a simple triplet form matrix}
}
\value{
a list of row pointer, column pointer, and values corresponding to a \link[Matrix:dsCMatrix-class]{Matrix::dsCMatrix} object
}
\description{
Convert a plain matrix or simple triplet form matrix to a \link[Matrix:dsCMatrix-class]{Matrix::dsCMatrix} (implicit, upper) form
}
