% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqp.R
\name{piqp_model}
\alias{piqp_model}
\title{The PIQP Solver Model Class}
\usage{
piqp_model(solver_ptr = NULL, dims = list(), dense_backend = logical(0))
}
\arguments{
\item{solver_ptr}{external pointer to the C++ solver object}

\item{dims}{a named list with elements \code{n}, \code{p}, and \code{m}}

\item{dense_backend}{logical flag indicating if the dense solver is used}
}
\description{
An S7 class wrapping the PIQP C++ Solver. Users will never
need to directly create instances of this class and should use the
more user-friendly functions \code{\link[=piqp]{piqp()}} and \code{\link[=solve_piqp]{solve_piqp()}}.
}
