% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{piqp_settings}
\alias{piqp_settings}
\title{Settings for PIQP}
\usage{
piqp_settings(
  rho_init = 1e-06,
  delta_init = 1e-04,
  eps_abs = 1e-08,
  eps_rel = 1e-09,
  check_duality_gap = TRUE,
  eps_duality_gap_abs = 1e-08,
  eps_duality_gap_rel = 1e-09,
  reg_lower_limit = 1e-10,
  reg_finetune_lower_limit = 1e-13,
  reg_finetune_primal_update_threshold = 7L,
  reg_finetune_dual_update_threshold = 5L,
  max_iter = 250L,
  max_factor_retires = 10L,
  preconditioner_scale_cost = FALSE,
  preconditioner_iter = 10L,
  tau = 0.99,
  iterative_refinement_always_enabled = FALSE,
  iterative_refinement_eps_abs = 1e-12,
  iterative_refinement_eps_rel = 1e-12,
  iterative_refinement_max_iter = 10L,
  iterative_refinement_min_improvement_rate = 5,
  iterative_refinement_static_regularization_eps = 1e-07,
  iterative_refinement_static_regularization_rel = .Machine$double.eps *
    .Machine$double.eps,
  verbose = FALSE,
  compute_timings = FALSE
)
}
\arguments{
\item{rho_init}{initial value for the primal proximal penalty parameter rho (`1e-6`)}

\item{delta_init}{initial value for the augmented lagrangian penalty parameter delta (`1e-4`)}

\item{eps_abs}{absolute tolerance (`1e-8`)}

\item{eps_rel}{relative tolerance (`1e-9`)}

\item{check_duality_gap}{check terminal criterion on duality gap (`TRUE`)}

\item{eps_duality_gap_abs}{absolute tolerance on duality gap (`1e-8`)}

\item{eps_duality_gap_rel}{relative tolerance on duality gap (`1e-9`)}

\item{reg_lower_limit}{lower limit for regularization (`1e-10`)}

\item{reg_finetune_lower_limit}{fine tune lower limit regularization (`1e-13`)}

\item{reg_finetune_primal_update_threshold}{threshold of number of no primal updates to transition to fine tune mode (`7L`)}

\item{reg_finetune_dual_update_threshold}{threshold of number of no dual updates to transition to fine tune mode (`5L`)}

\item{max_iter}{maximum number of iterations (`250L`)}

\item{max_factor_retires}{maximum number of factorization retires before failure (`10L`)}

\item{preconditioner_scale_cost}{scale cost in Ruiz preconditioner (`FALSE`)}

\item{preconditioner_iter}{maximum of preconditioner iterations (`10L`)}

\item{tau}{maximum interior point step length (`0.99`)}

\item{iterative_refinement_always_enabled}{always run iterative refinement and not only on factorization failure (`FALSE`)}

\item{iterative_refinement_eps_abs}{iterative refinement absolute tolerance (`1e-12`)}

\item{iterative_refinement_eps_rel}{iterative refinement relative tolerance (`1e-12`)}

\item{iterative_refinement_max_iter}{maximum number of iterations for iterative refinement (`10L`)}

\item{iterative_refinement_min_improvement_rate}{minimum improvement rate for iterative refinement (`5.0`)}

\item{iterative_refinement_static_regularization_eps}{static regularization for KKT system for iterative refinement (`1e-7`)}

\item{iterative_refinement_static_regularization_rel}{static regularization w.r.t. the maximum abs diagonal term of KKT system (`.Machine$double.eps * .Machine$double.eps`)}

\item{verbose}{verbose printing (`FALSE`)}

\item{compute_timings}{measure timing information internally (`FALSE`)}
}
\description{
For further details please consult the PIQP documentation:
\url{https://predict-epfl.github.io/piqp/}
}
