% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piqp_sparse_solve}
\alias{piqp_sparse_solve}
\title{Solve using the piqp sparse matrix interface}
\usage{
piqp_sparse_solve(P, c, A, b, G, h, x_lb, x_ub, rs)
}
\arguments{
\item{P}{a symmetric positive definite matrix}

\item{c}{the objective cost}

\item{A}{the coefficient matrix of primal constraints}

\item{b}{the right hand side vector of the primal constraints}

\item{G}{the matrix of inequality constraints}

\item{h}{the right side vector of the matrix inequality constraints}

\item{x_lb}{the lower bounds for the solution}

\item{x_ub}{the upper bounds for the solution}

\item{rs}{a list of control settings}
}
\value{
named list
\itemize{
\item \code{x} (primal solution)
\item \code{y} (dual solution of equality constraints)
\item \code{z} (dual solution of inequality constraints)
\item \code{z_lb} (dual solution of lower bound box constraints)
\item \code{z_ub} (dual solution of upper bound box constraints)
\item \code{obj} (primal objective value)
\item \code{run_time} (total runtime, if so specified in settings)
}
}
\description{
Solve using the piqp sparse matrix interface
}
