% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqp.R
\name{update_settings}
\alias{update_settings}
\title{Update settings of a PIQP model}
\usage{
update_settings(model, ...)
}
\arguments{
\item{model}{A \code{piqp_model} object}

\item{...}{not used}
}
\value{
invisible NULL
}
\description{
Update settings of a PIQP model
}
