% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqp.R
\name{piqp}
\alias{piqp}
\title{Interface to 'piqp', a proximal interior point quadratic solver}
\usage{
piqp(
  A = matrix(0, nrow = 0, ncol = 0),
  b = numeric(0),
  c = numeric(0),
  P = matrix(0, nrow = 0, ncol = 0),
  G = matrix(0, nrow = 0, ncol = 0),
  h = numeric(0),
  x_lb = NULL,
  x_ub = NULL,
  control = list(),
  interface = c("auto", "dense", "sparse")
)
}
\arguments{
\item{A}{a matrix of constraint coefficients}

\item{b}{a numeric vector giving the primal constraints}

\item{c}{a numeric vector giving the primal objective}

\item{P}{a symmetric positive semidefinite matrix}

\item{G}{a matrix of inequality coefficients}

\item{h}{a numeric vector of the inequality right hand side}

\item{x_lb}{a numeric vector of upper bounds, default \code{NULL}
indicating \code{Inf} for all variables, otherwise should be number of
variables long}

\item{control}{a list giving specific control parameters to use in
place of default values, with empty list indicating defaults}

\item{interface}{whether to use sparse, or dense interface. Default
is "auto" mode will switch to sparse interface if any of the
specified matrices are sparse as in the Matrix package.}
}
\value{
named list
\itemize{
\item \code{status} (status code)
\item \code{x} (primal solution)
\item \code{y} (dual solution of equality constraints)
\item \code{z} (dual solution of inequality constraints)
\item \code{z_lb} (dual solution of lower bound box constraints)
\item \code{z_ub} (dual solution of upper bound box constraints)
\item \code{pobj} (primal objective value)
\item \code{dobj} (dual objective value)
\item \code{dgap} (absolute duality gap absolute)
\item \code{dgap_rel} (relative duality gap)
\item \code{run_time} (total runtime, if so specified in settings)
}
}
\description{
PIQP solves linear programs (LPs) and quadratic programs (QPs). ,

Minimize \deqn{\frac{1}{2}x^TPx + c^Tx} subject to \deqn{Ax = b},
\deqn{Gx \leq h}, \deqn{x_{lb} \leq x \leq x_{ub}}, where \eqn{x
  \in R^n}, \eqn{h \in R^m}, \eqn{P = P^T} and
nonnegative-definite, \eqn{c \in R^n}, \eqn{A \in R^{p\times n}},
\eqn{G \in R^{m\times n}}, \eqn{b \in R^p}, \eqn{x{lb} \in R^n},
\eqn{x_{ub} \in R^n}.
}
\details{
PIQP uses compressed column sparse (CSC) format
internally, which is the \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix} in the Matrix
package. So even P should be in CSC format, not taking advantage
of symmetry for instance. This package accepts either
\link[Matrix:dgCMatrix-class]{Matrix::dgcMatrix-class} class objects or the triplet format sparse
matrices as in the \code{\link[slam:matrix]{slam::simple_triplet_matrix()}}
}
\seealso{
\code{\link[=piqp_control]{piqp_control()}}, \code{\link[=status_descriptions]{status_descriptions()}}
}
